split parsing and utils tests
