+++ config_utilities/test/CMakeLists.txt
@@ -29,7 +29,8 @@ add_executable(
   tests/update.cpp
   tests/validity_checks.cpp
   tests/virtual_config.cpp
-  tests/yaml_parsing.cpp)
+  tests/yaml_parsing.cpp
+  tests/yaml_utils.cpp)
 target_include_directories(test_${PROJECT_NAME} PRIVATE include)
 target_link_libraries(test_${PROJECT_NAME} PRIVATE ${PROJECT_NAME}
                                                    GTest::gtest_main)
